Implement basic movement and fix project structure
Implement a basic movement system that can allow the player to move left, right, and jump. If the player is airborne, disable the jump ability. Also add a physics material to the box collider to avoid the sticky walls situation.

Also fix the project structure to implement folders for scripts, materials, and sprites.